+++ include/contraction/ch_edge.hpp
@@ -1,5 +1,5 @@
 /*PGR-GNU*****************************************************************
-File: ch_graphs.hpp
+File: ch_edge.hpp
 
 Generated with Template by:
 Copyright (c) 2015 pgRouting developers
@@ -27,36 +27,46 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
  ********************************************************************PGR-GNU*/
 
-#ifndef INCLUDE_CONTRACTION_CH_GRAPHS_HPP_
-#define INCLUDE_CONTRACTION_CH_GRAPHS_HPP_
+#ifndef INCLUDE_CPP_COMMON_CH_EDGE_HPP_
+#define INCLUDE_CPP_COMMON_CH_EDGE_HPP_
 #pragma once
 
-
-#include <limits>
-#include <algorithm>
-#include <vector>
-
-#include "contraction/contractionGraph.hpp"
-#include "cpp_common/base_graph.hpp"
-#include "cpp_common/ch_vertex.hpp"
-#include "cpp_common/ch_edge.hpp"
-
+#include <iostream>
+#include <sstream>
+#include "contraction/ch_vertex.hpp"
 
 namespace pgrouting {
-namespace graph {
 
-using CHUndirectedGraph =  Pgr_contractionGraph <
-    boost::adjacency_list < boost::listS, boost::vecS,
-    boost::undirectedS,
-    CH_vertex, CH_edge>, false>;
+class CH_edge {
+ public:
+     CH_edge() = default;
 
-using CHDirectedGraph = Pgr_contractionGraph<
-    boost::adjacency_list < boost::listS, boost::vecS,
-    boost::bidirectionalS,
-    CH_vertex, CH_edge>, true>;
+     CH_edge(int64_t eid, int64_t source, int64_t target, double cost) :
+         id(eid), source(source),
+         target(target), cost(cost) {}
 
+     void cp_members(const CH_edge &other);
 
-}  // namespace graph
+     void add_contracted_vertex(CH_vertex& v);
+     void add_contracted_edge_vertices(CH_edge& e);
+
+     bool has_contracted_vertices() const;
+
+     void clear_contracted_vertices() {m_contracted_vertices.clear();}
+     const Identifiers<int64_t>& contracted_vertices() const;
+     Identifiers<int64_t>& contracted_vertices();
+     friend std::ostream& operator <<(std::ostream& os, const CH_edge& e);
+
+ public:
+     int64_t id;
+     int64_t source;
+     int64_t target;
+     double cost;
+
+ private:
+     Identifiers<int64_t> m_contracted_vertices;
+};
+
 }  // namespace pgrouting
 
-#endif  // INCLUDE_CONTRACTION_CH_GRAPHS_HPP_
+#endif  // INCLUDE_CPP_COMMON_CH_EDGE_HPP_
