+++ include/contraction/ch_vertex.hpp
