+++ src/contraction/ch_edge.cpp
@@ -27,7 +27,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
  ********************************************************************PGR-GNU*/
 
-#include "cpp_common/ch_edge.hpp"
+#include "contraction/ch_edge.hpp"
 
 namespace pgrouting {
